day04/pyproject.toml
[project]
name = "day04"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.80",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
